Added more tests
Reorganized libtiny_tui tests
